+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorSettings.h
@@ -316,57 +316,89 @@ private:
 
 private:
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Assembly name"))
-		FString AssemblyName;
+	FString AssemblyName;
 
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Deployment name"))
-		FString PrimaryDeploymentName;
+	FString PrimaryDeploymentName;
 
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Cloud launch configuration path"))
-		FFilePath PrimaryLaunchConfigPath;
+	FFilePath PrimaryLaunchConfigPath;
 
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Snapshot path"))
-		FFilePath SnapshotPath;
+	FFilePath SnapshotPath;
 
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Region"))
-		TEnumAsByte<ERegionCode::Type> PrimaryDeploymentRegionCode;
+	TEnumAsByte<ERegionCode::Type> PrimaryDeploymentRegionCode;
 
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Main Deployment Cluster"))
-		FString MainDeploymentCluster;
+	FString MainDeploymentCluster;
 
 	/** Tags used when launching a deployment */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Deployment tags"))
-		FString DeploymentTags;
+	FString DeploymentTags;
 
 	const FString SimulatedPlayerLaunchConfigPath;
 
 public:
+	/** The build configuration to use when creating workers for the assembly, e.g. Development */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly", meta = (DisplayName = "Build Configuration"))
+	FString AssemblyBuildConfiguration;
+
+	/** Allow overwriting an assembly of the same name */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly", meta = (DisplayName = "Force Assembly Overwrite"))
+	bool bForceAssemblyOverwrite;
+
+	/** Whether to build client worker as part of the assembly */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly", meta = (DisplayName = "Build Client Worker"))
+	bool bBuildClientWorker;
+
+	/** Whether to generate schema automatically before building an assembly */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly", meta = (DisplayName = "Generate Schema"))
+	bool bGenerateSchema;
+
+	/** Whether to generate a snapshot automatically before building an assembly */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly", meta = (DisplayName = "Generate Snapshot"))
+	bool bGenerateSnapshot;
+
+	/** Extra arguments to pass when building the server worker. */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly")
+	FString BuildServerExtraArgs;
+
+	/** Extra arguments to pass when building the client worker. */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly")
+	FString BuildClientExtraArgs;
+
+	/** Extra arguments to pass when building the simulated player worker. */
+	UPROPERTY(EditAnywhere, config, Category = "Assembly")
+	FString BuildSimulatedPlayerExtraArgs;
+
 	/** If the Development Authentication Flow is used, the client will try to connect to the cloud rather than local deployment. */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-		bool bUseDevelopmentAuthenticationFlow;
+	bool bUseDevelopmentAuthenticationFlow;
 
 	/** The token created using 'spatial project auth dev-auth-token' */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-		FString DevelopmentAuthenticationToken;
+	FString DevelopmentAuthenticationToken;
 
 	/** The deployment to connect to when using the Development Authentication Flow. If left empty, it uses the first available one (order not guaranteed when there are multiple items). The deployment needs to be tagged with 'dev_login'. */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-		FString DevelopmentDeploymentToConnect;
+	FString DevelopmentDeploymentToConnect;
 
 private:
 	UPROPERTY(EditAnywhere, config, Category = "Simulated Players", meta = (EditCondition = "bSimulatedPlayersIsEnabled", DisplayName = "Region"))
-		TEnumAsByte<ERegionCode::Type> SimulatedPlayerDeploymentRegionCode;
+	TEnumAsByte<ERegionCode::Type> SimulatedPlayerDeploymentRegionCode;
 
 	UPROPERTY(EditAnywhere, config, Category = "Cloud", meta = (DisplayName = "Simulated Player Cluster"))
-		FString SimulatedPlayerCluster;
+	FString SimulatedPlayerCluster;
 
 	UPROPERTY(EditAnywhere, config, Category = "Simulated Players", meta = (DisplayName = "Include simulated players"))
-		bool bSimulatedPlayersIsEnabled;
+	bool bSimulatedPlayersIsEnabled;
 
 	UPROPERTY(EditAnywhere, config, Category = "Simulated Players", meta = (EditCondition = "bSimulatedPlayersIsEnabled", DisplayName = "Deployment name"))
-		FString SimulatedPlayerDeploymentName;
+	FString SimulatedPlayerDeploymentName;
 
 	UPROPERTY(EditAnywhere, config, Category = "Simulated Players", meta = (EditCondition = "bSimulatedPlayersIsEnabled", DisplayName = "Number of simulated players"))
-		uint32 NumberOfSimulatedPlayers;
+	uint32 NumberOfSimulatedPlayers;
 
 	static bool IsRegionCodeValid(const ERegionCode::Type RegionCode);
 	static bool IsManualWorkerConnectionSet(const FString& LaunchConfigPath, TArray<FString>& OutWorkersManuallyLaunched);
@@ -516,6 +548,12 @@ public:
 		return DeploymentTags;
 	}
 
+	void SetAssemblyBuildConfiguration(const FString& Configuration);
+	FORCEINLINE FText GetAssemblyBuildConfiguration() const
+	{
+		return FText::FromString(AssemblyBuildConfiguration);
+	}
+
 	void SetSimulatedPlayerRegionCode(const ERegionCode::Type RegionCode);
 	FORCEINLINE FText GetSimulatedPlayerRegionCode() const
 	{
@@ -533,6 +571,30 @@ public:
 	FORCEINLINE bool IsSimulatedPlayersEnabled() const
 	{
 		return bSimulatedPlayersIsEnabled;
+	}
+
+	void SetForceAssemblyOverwrite(bool bForce);
+	FORCEINLINE bool IsForceAssemblyOverwriteEnabled() const
+	{
+		return bForceAssemblyOverwrite;
+	}
+
+	void SetBuildClientWorker(bool bBuild);
+	FORCEINLINE bool IsBuildClientWorkerEnabled() const
+	{
+		return bBuildClientWorker;
+	}
+
+	void SetGenerateSchema(bool bGenerate);
+	FORCEINLINE bool IsGenerateSchemaEnabled() const
+	{
+		return bGenerateSchema;
+	}
+
+	void SetGenerateSnapshot(bool bGenerate);
+	FORCEINLINE bool IsGenerateSnapshotEnabled() const
+	{
+		return bGenerateSnapshot;
 	}
 
 	void SetUseGDKPinnedRuntimeVersion(bool IsEnabled);
@@ -565,7 +627,7 @@ public:
 	}
 
 	void SetNumberOfSimulatedPlayers(uint32 Number);
-	FORCEINLINE uint32 GetNumberOfSimulatedPlayer() const
+	FORCEINLINE uint32 GetNumberOfSimulatedPlayers() const
 	{
 		return NumberOfSimulatedPlayers;
 	}